♻️ extract unpaid merged pr count logic
