+++ frontend/src/pages/ProjectDetails/Payments/PaymentForm/ContributorSelect/index.tsx
@@ -5,15 +5,12 @@ import {
   GetProjectContributorsForPaymentSelectDocument,
   GetProjectContributorsForPaymentSelectQuery,
   GithubContributorFragment,
-  ProjectContributorsForPaymentSelectFragment,
-  PullDetailsFragment,
   SearchGithubUsersByHandleSubstringDocument,
   SearchGithubUsersByHandleSubstringQuery,
 } from "src/__generated/graphql";
 import { getContributors } from "src/utils/project";
 import View from "./View";
 import { useLocation } from "react-router-dom";
-import { chain, flatMap, some } from "lodash";
 
 type Props = {
   projectId: string;
@@ -32,11 +29,12 @@ export default function ContributorSelect({ projectId, contributor, setContribut
 
   const getProjectContributorsQuery = useHasuraQuery<GetProjectContributorsForPaymentSelectQuery>(
     GetProjectContributorsForPaymentSelectDocument,
-    HasuraUserRole.Public,
+    HasuraUserRole.RegisteredUser,
     {
       variables: { projectId },
     }
   );
+
   const searchGithubUsersByHandleSubstringQuery = useHasuraQuery<SearchGithubUsersByHandleSubstringQuery>(
     SearchGithubUsersByHandleSubstringDocument,
     HasuraUserRole.RegisteredUser,
@@ -46,22 +44,10 @@ export default function ContributorSelect({ projectId, contributor, setContribut
     }
   );
 
-  const unpaidMergedPullsCountByContributor = useMemo(
-    () =>
-      getProjectContributorsQuery.data?.projectsByPk
-        ? countUnpaidMergedPullsByContributor(getProjectContributorsQuery.data?.projectsByPk)
-        : {},
+  const { contributors: internalContributors } = useMemo(
+    () => getContributors(getProjectContributorsQuery.data?.projectsByPk),
     [getProjectContributorsQuery.data]
   );
-
-  const { contributors: internalContributors } = useMemo(() => {
-    const contributors = getContributors(getProjectContributorsQuery.data?.projectsByPk);
-    contributors.contributors.map(c => ({
-      ...c,
-      unpaidMergedPullsCount: unpaidMergedPullsCountByContributor[c?.login],
-    }));
-    return contributors;
-  }, [getProjectContributorsQuery.data, unpaidMergedPullsCountByContributor]);
 
   const filteredContributors = sortListByLogin(
     internalContributors.filter(
@@ -112,24 +98,3 @@ function sortListByLogin<T extends { login: string }>(objectsWithLogin: T[] | nu
       )
     : [];
 }
-
-export const countUnpaidMergedPullsByContributor = ({
-  githubRepos,
-  budgets,
-}: ProjectContributorsForPaymentSelectFragment) => {
-  const paidItemsByLogin = chain(budgets)
-    .flatMap("paymentRequests")
-    .groupBy("githubRecipient.login")
-    .mapValues(requests => flatMap(requests, "workItems"))
-    .value();
-
-  const notPaid = ({ author, repoId, number: issueNumber }: PullDetailsFragment) =>
-    !some(paidItemsByLogin[author.login], { repoId, issueNumber });
-
-  return chain(githubRepos)
-    .flatMap("githubRepoDetails.pullRequests")
-    .filter("mergedAt")
-    .filter(notPaid)
-    .countBy("author.login")
-    .value();
-};
